refactor(domain): Models alterados para permitir a criação via builder pattern
